Make graph weights uniform cost during traversal to prevent too many packs being used
If the actual weights are used then we'll traverse a path which favours
the "shorter" distance, which usually means a greater number of hops across
low weight edges. In this case we prefer larger "distances" as pack sizes
are used as weight. This feels like a bit of a hack but I couldn't get
the heuristics to favour higher weight edges.